requirements.txt
django-spirit
-e git+https://github.com/vsudilov/Spirit.git#egg=django-spirit
